chore: rename "static_file" to "file_server"
