+++ src/addon/file_server/scoped_file_system.rs
